Feedback when parsing fails
This will not print the token value of the token that causes the parsing
to fail. We track this in builder.ErrorToken which was added to rd.

rd was internalized into internal/rd to make this easier. This still
doesn't print line numbers, but it's a start. Maybe I'll fix this in
cffmt2.

Signed-off-by: Miek Gieben <miek@miek.nl>