Add NamespaceFunction
This match 'xxx:yyyy' as a valid function. The tokens we get are
"strange": KeywordDeclaration Punctuation NameFunction

grab them all and add them to the AST. Then when printing we loose the
":" (Punctuation) so just print that too, but special case it (hack!) to
not also print otherones, like "," and "}" which handle in the correct
places.

Add lexer support for this was not working, so just parse it like this.

Closes: #36

Signed-off-by: Miek Gieben <miek@miek.nl>